snapshot: Move compress functions to separate file
