+++ src/test/csrc/snapshot.h
@@ -35,8 +35,6 @@ public:
     this->size += size;
   }
 
-  long compressToFile(uint8_t *ptr, const char *filename, long buf_size);
-
   void close() { }
   void flush();
   void save();
@@ -64,12 +62,6 @@ public:
   void open(const std::string& filename) VL_MT_UNSAFE_ONE { open(filename.c_str()); }
 
   long unbuf_read(uint8_t* dest, long rsize);
-
-  // Return whether the file is a gz file
-  int isGzFile(const char *filename);
-
-  // Read binary from .gz file
-  long readFromGz(void* ptr, const char *file_name, long buf_size);
 
   void close() override VL_MT_UNSAFE_ONE;
   void flush() override VL_MT_UNSAFE_ONE {}
